requirements.txt
Django==2.2.2
djangorestframework==3.9.4
mysqlclient==1.4.2.post1
requests[security]==2.22.0
